WIP: Dynamically create test data
Reimplement CSharpDefinitionFormatter test cases for types using dynamically generated assembly.
Remove old CSharpDefinitionFormatterTest class
