+++ src/MdDocs.ApiReference.Model.Test/ModuleDocumentationTest.cs
@@ -45,35 +45,6 @@ namespace Grynwald.MdDocs.ApiReference.Test.Model
                 typeof(TestClass_InterfaceImplementation),
                 typeof(TestStruct_InterfaceImplementation),
                 typeof(TestInterface_Inheritance),
-                typeof(TestClass_CSharpDefinition),
-                typeof(CSharpDefinitionTest1Attribute),
-                typeof(CSharpDefinitionTest2Attribute),
-                typeof(CSharpDefinitionTest3Attribute),
-                typeof(CSharpDefinitionTest4Attribute),               
-                typeof(CSharpDefinitionTestFlagsEnum),
-                typeof(CSharpDefinitionTest_StaticClass),
-                typeof(CSharpDefinitionTest_AbstractClass),
-                typeof(CSharpDefinitionTest_SealedClass),
-                typeof(CSharpDefinitionTest_GenericClass<>),
-                typeof(CSharpDefinitionTest_GenericClass2<,>),
-                typeof(CSharpDefinitionTest_ExtensionClass),
-                typeof(CSharpDefinitionTest_ClassWithAttributes),
-                typeof(CSharpDefinitionTest_ClassWithInterfaceImplementation),
-                typeof(CSharpDefinitionTest_ClassWithMultipleInterfaceImplementations),
-                typeof(CSharpDefinitionTest_ClassWithbaseTypeAndInterfaceImplementations),
-                typeof(CSharpDefinitionTest_ClassWithBaseType),
-                typeof(CSharpDefinitionTest_Interface),
-                typeof(CSharpDefinitionTest_InterfaceWithMultipleInterfaceImplementations),
-                typeof(CSharpDefinitionTest_Struct),
-                typeof(CSharpDefinitionTest_StructWithMultipleInterfaceImplementations),
-                typeof(CSharpDefinitionTest_ReadOnlyStruct),
-                typeof(CSharpDefinitionTest_GenericInterface_Contravariant<>),
-                typeof(CSharpDefinitionTest_GenericInterface_Covariant<>),
-                typeof(CSharpDefinitionTestFlagsEnum2),
-                typeof(CSharpDefinitionTest_ClassWithAttribute2),
-                typeof(CSharpDefinitionTest_ClassWithAttribute3),
-                typeof(CSharpDefinitionTest6Attribute),
-                typeof(CSharpDefinitionTestEnum),
                 typeof(TestClass_XmlDocs<,>),
                 typeof(TestClass_NoDocumentation),
                 typeof(TestClass_RefParameters),
